PR #1261: Updated scalar logic and tidied up tests
